fix: 로그인 시 토큰 보내는 방식 수정

+++ packages/api/manageProduct/deleteProduct/api.ts
@@ -1,15 +1,19 @@
 import { instance } from "../../instance/instance";
+import Cookies from "js-cookie";
 
 export const deleteProduct = async (productId) => {
-  // let token = Cookies.get("token");
+  const token = Cookies.get("token");
+
+  if (!token) {
+    throw new Error("No authentication token found");
+  }
+
   const { data } = await instance.delete(`/products/${productId}`, {
     headers: {
-      // 'Authorization': `Bearer ${token}
-      'Authorization': `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJzdWIiOiIwIiwiaWF0IjoxNzAyNDc1Njg0LCJleHAiOjE3MDI4MzU2ODR9.LF_zLUg36jorKfkGrdIFTOgF64naMk9ZBmAJMzCBWFU`
-    }
+      Authorization: `Bearer ${token}`,
+    },
   });
   console.log(data);
-  // console.log(token);
   window.location.reload();
   return data;
 };
